+++ core/mysettings.py
@@ -1,6 +1,7 @@
-from PyQt4.QtGui import QColor
 
-from ..qgissettingmanager import *
+from PyQt4.QtCore import Qt
+from PyQt4.QtGui import QColor
+from linkit.qgissettingmanager import *
 
 pluginName = "linkit"
 
@@ -9,6 +10,8 @@ class MySettings(SettingManager):
     def __init__(self):
         SettingManager.__init__(self, pluginName)
         self.addSetting("drawButton", "bool", "global", True)
-        self.addSetting("dockArea", "integer", "global", 0)
         self.addSetting("rubberColor", "Color", "global", QColor(0, 0, 255, 150), {"alpha": True})
         self.addSetting("rubberWidth", "double", "global", 2)
+        self.addSetting("dockVisible", "bool", "global", False)
+        self.addSetting("dockArea", "integer", "global", int(Qt.LeftDockWidgetArea))
+        self.addSetting("drawEnabled", "bool", "global", False)