+++ gui/linkerdock.py
@@ -6,251 +6,236 @@ Link It
 QGIS module
 """
 
-from PyQt4.QtCore import pyqtSlot, Qt
-from PyQt4.QtGui import QDockWidget, QIcon
-from qgis.core import QgsFeature, QgsFeatureRequest, QgsPoint, QgsGeometry, QGis, QgsMapLayerRegistry
-from qgis.gui import QgsRubberBand, QgsMessageBar
-from math import pi, sqrt, sin, cos
+from PyQt4.QtCore import pyqtSlot, Qt, QSettings, QTimer
+from PyQt4.QtGui import QDockWidget, QColor
+from qgis.core import QgsProject, QgsFeature, QgsRelation, QGis
+from qgis.gui import QgsRubberBand, QgsMessageBar, QgsEditorWidgetRegistry, QgsMapToolIdentifyFeature, QgsAttributeEditorContext, QgsHighlight
 
-from ..qgissettingmanager import SettingDialog
-from ..core.mysettings import MySettings
-from ..ui.ui_linker import Ui_linker
-from maptoolgetfeature import MapToolGetFeature
-
-
-def floatrange(a, b, step):
-    while a < b:
-        yield a
-        a += step
-    yield b
-
-
-def centroid(feature):
-    geom = feature.geometry()
-    if geom.type() == QGis.Line:
-        return geom.interpolate(geom.length()/2).asPoint()
-    else:
-        return geom.centroid().asPoint()
-
-
-def castFeatureId(text):
-    try:
-        return long(text)
-    except ValueError:
-        return None
+from linkit.qgissettingmanager import SettingDialog
+from linkit.core.mysettings import MySettings
+from linkit.core.arc import arc
+from linkit.ui.ui_linker import Ui_linker
 
 
 class LinkerDock(QDockWidget, Ui_linker, SettingDialog):
     def __init__(self, iface):
-        self.destinationLayer = None
-        self.destinationProvider = None
-        self.destinationField = None
-        self.sourceLayer = None
-        self.featureId = None
+        # QGIS
         self.iface = iface
-        self.mapCanvas = iface.mapCanvas()
         self.settings = MySettings()
-        self.linkRubber = QgsRubberBand(self.mapCanvas)
-        self.featureRubber = QgsRubberBand(self.mapCanvas)
-        self.mapTool = None
+        self.linkRubber = QgsRubberBand(self.iface.mapCanvas())
+        self.featureHighlight = None
+        # Relation management
+        self.relationManager = QgsProject.instance().relationManager()
+        self.relationManager.changed.connect(self.loadRelations)
+        self.relation = QgsRelation()
+        self.referencingFeature = QgsFeature()
+        self.relationWidgetWrapper = None
+        self.editorContext = QgsAttributeEditorContext()
+        self.editorContext.setVectorLayerTools(self.iface.vectorLayerTools())
 
+        # GUI
         QDockWidget.__init__(self)
         self.setupUi(self)
         SettingDialog.__init__(self, MySettings(), False, True)
+        self.drawButton.setChecked(self.settings.value("drawEnabled"))
 
-    def set(self, destinationLayerId, destinationField, sourceLayerId, featureId):
-        self.disconnectLayers()
+        self.relationReferenceWidget.setAllowMapIdentification(True)
+        self.relationReferenceWidget.setEmbedForm(False)
 
-        self.destinationLayer = QgsMapLayerRegistry.instance().mapLayer(destinationLayerId)
-        self.destinationField = destinationField
-        self.sourceLayer = QgsMapLayerRegistry.instance().mapLayer(sourceLayerId)
+        self.mapTool = QgsMapToolIdentifyFeature(self.iface.mapCanvas())
+        self.mapTool.setButton(self.identifyReferencingFeatureButton)
 
-        self.featureIdLabel.setText("%u" % featureId)
-        self.featureId = featureId
+        # Connect signal/slot
+        self.relationComboBox.currentIndexChanged.connect(self.currentRelationChanged)
+        self.mapTool.featureIdentified.connect(self.setReferencingFeature)
 
-        self.enableUI()
+        # load relations at start
+        self.loadRelations()
 
-        self.featureRubber.reset()
-
-        if self.mapTool is not None:
-            self.mapCanvas.unsetMapTool(self.mapTool)
-        
-        self.cancelButton.setIcon(QIcon(":/plugins/linkit/icons/cancel.svg"))
-        self.deleteButton.setIcon(QIcon(":/plugins/linkit/icons/delete.svg"))
-        self.drawButton.setIcon(QIcon(":/plugins/linkit/icons/drawline.svg"))
-        self.selectButton.setIcon(QIcon(":/plugins/linkit/icons/maptool.svg"))
-
-        try:
-            self.sourceLayer.layerDeleted.connect(self.close)
-        except:
-            pass
-        try:
-            self.destinationLayer.layerDeleted.connect(self.close)
-            self.destinationLayer.editingStarted.connect(self.enableUI)
-            self.destinationLayer.editingStopped.connect(self.enableUI)
-            self.destinationLayer.featureDeleted.connect(self.featureDeleted)
-        except:
-            pass
-
-        self.show()
-
-    def disconnectLayers(self):
-        try:
-            self.destinationLayer.layerDeleted.disconnect(self.close)
-            self.destinationLayer.editingStarted.disconnect(self.enableUI)
-            self.destinationLayer.editingStopped.disconnect(self.enableUI)
-            self.destinationLayer = None
-        except:
-            pass
-        try:
-            self.sourceLayer.layerDeleted.disconnect(self.enableUI)
-            self.sourceLayer = None
-        except:
-            pass
-
-    def enableUI(self):
-        self.mapCanvas.unsetMapTool(self.mapTool)
-
-        enable = False
-
-        feature = self.getFeature()
-        ok = feature[0]
-
-        if not ok:
-            self.iface.messageBar().pushMessage("Link It", feature[1], QgsMessageBar.WARNING, 4)
-        else:
-            # update current value
-            currentValue = feature[1][self.destinationField]
-            self.linkedItemID.setText("%s" % currentValue)
-            self.drawLink()
-
-        # destination layer is editable
-        if ok and self.destinationLayer.isEditable():
-            enable = True
-
-        # enable UI
-        self.editButtonWidget.setEnabled(enable)
-        self.selectButton.show()
-        self.cancelButton.hide()
-
-    def getFeature(self):
-        # check layers
-        if self.destinationLayer is None:
-            return False, "Destination layer does not exist."
-
-        if self.sourceLayer is None:
-            return False, "Source layer does not exist."
-
-        # check feature exist and fetch feature at once
-        feature = QgsFeature()
-        if self.destinationLayer.getFeatures(QgsFeatureRequest().setFilterFid(self.featureId)).nextFeature(feature) is False:
-            return False, "Could not fetch feature at ID %u." % self.featureId
-
-        # check destination field exist
-        if self.destinationLayer.fieldNameIndex(self.destinationField) == -1:
-            return False, "Destination field (%s) does not exist." % self.featureId
-
-        return True, feature
-
-    def featureDeleted(self, fid):
-        if self.featureId is None:
-            return
-        if self.featureId == fid:
-            self.close()
-
-    def closeEvent(self, e):
-        if self.mapTool is not None:
-            self.mapCanvas.unsetMapTool(self.mapTool)
-        self.linkRubber.reset()
-        self.featureRubber.reset()
-        self.disconnectLayers()
-
-    @pyqtSlot(name="on_selectButton_clicked")
-    def setMapTool(self):
-        self.mapTool = MapToolGetFeature(self.mapCanvas, self.sourceLayer)
-        self.mapTool.featureIdentified.connect(self.featureIdentified)
-        self.mapCanvas.setMapTool(self.mapTool)
-        self.selectButton.hide()
-        self.cancelButton.show()
-
-    @pyqtSlot(name="on_cancelButton_clicked")
-    def unsetMapTool(self, dummy=None):
-        self.mapCanvas.unsetMapTool(self.mapTool)
-        self.selectButton.show()
-        self.cancelButton.hide()
-
-    @pyqtSlot(name="on_deleteButton_clicked")
-    def deleteLink(self):
-        self.linkedItemID.clear()
-        self.updateLink("")
-
-    def updateLink(self, new):
-        new = castFeatureId(new)
-        fldIdx = self.destinationLayer.fieldNameIndex(self.destinationField)
-        if self.destinationLayer.isEditable():
-            if not self.destinationLayer.editBuffer().changeAttributeValue(self.featureId, fldIdx, new):
-                self.iface.messageBar().pushMessage("Link It", "Could not update link value.",
-                                                    QgsMessageBar.WARNING, 4)
-        else:
-            self.iface.messageBar().pushMessage("Link It", "Destination layer must be editable.",
-                                                QgsMessageBar.WARNING, 4)
+    def showEvent(self, QShowEvent):
         self.drawLink()
 
-    def featureIdentified(self, new):
-        self.linkedItemID.setText("%s" % new)
-        self.updateLink(new)
-        self.unsetMapTool()
+    def closeEvent(self, e):
+        self.iface.mapCanvas().unsetMapTool(self.mapTool)
+        self.linkRubber.reset()
+        self.deleteHighlight()
+        if self.relation.isValid():
+            self.relation.referencingLayer().editingStarted.disconnect(self.relationEditableChanged)
+            self.relation.referencingLayer().editingStopped.disconnect(self.relationEditableChanged)
+            self.relation.referencingLayer().attributeValueChanged.disconnect(self.layerValueChangedOutside)
+
+    def runForFeature(self, relationId, layer, feature):
+        index = self.relationComboBox.findData(relationId)
+        self.relationComboBox.setCurrentIndex(index)
+        self.setReferencingFeature(feature)
+        self.show()
+        if not layer.isEditable():
+            self.iface.messageBar().pushMessage("Link It", "Cannot set a new related feature since %s is not editable" % layer.name(), QgsMessageBar.WARNING, 4)
+        else:
+            self.relationReferenceWidget.mapIdentification()
+
+    @pyqtSlot(name="on_identifyReferencingFeatureButton_clicked")
+    def activateMapTool(self):
+        self.iface.mapCanvas().setMapTool(self.mapTool)
+
+    def deactivateMapTool(self):
+        self.iface.mapCanvas().unsetMapTool(self.mapTool)
+
+    def loadRelations(self):
+        self.relation = QgsRelation()
+        self.referencingFeature = QgsFeature()
+        self.relationComboBox.currentIndexChanged.disconnect(self.currentRelationChanged)
+        self.relationComboBox.clear()
+        for relation in self.relationManager.referencedRelations():
+            if relation.referencingLayer().hasGeometryType():
+                self.relationComboBox.addItem(relation.name(), relation.id())
+        self.relationComboBox.setCurrentIndex(0)
+        self.relationComboBox.currentIndexChanged.connect(self.currentRelationChanged)
+        self.currentRelationChanged(0)
+
+    def currentRelationChanged(self, index):
+        # disconnect previous relation
+        if self.relation.isValid():
+            self.relation.referencingLayer().editingStarted.disconnect(self.relationEditableChanged)
+            self.relation.referencingLayer().editingStopped.disconnect(self.relationEditableChanged)
+            self.relation.referencingLayer().attributeValueChanged.disconnect(self.layerValueChangedOutside)
+
+        self.referencingFeatureLayout.setEnabled(index >= 0)
+        relationId = self.relationComboBox.itemData(index)
+        self.relation = self.relationManager.relation(relationId)
+        self.mapTool.setLayer(self.relation.referencingLayer())
+        self.setReferencingFeature()
+        # connect
+        if self.relation.isValid():
+            self.relation.referencingLayer().editingStarted.connect(self.relationEditableChanged)
+            self.relation.referencingLayer().editingStopped.connect(self.relationEditableChanged)
+            self.relation.referencingLayer().attributeValueChanged.connect(self.layerValueChangedOutside)
+
+    def setReferencingFeature(self, feature=QgsFeature()):
+        self.deactivateMapTool()
+        self.referencingFeature = QgsFeature(feature)
+
+        del self.relationWidgetWrapper
+
+        # disable relation reference widget if no referencing feature
+        self.referencedFeatureLayout.setEnabled(feature.isValid())
+
+        # set line edit
+        if not self.relation.isValid() or not feature.isValid():
+            self.referencingFeatureLineEdit.clear()
+            self.relationReferenceWidget.setForeignKey(None)
+            self.relationWidgetWrapper = None
+            return
+        self.referencingFeatureLineEdit.setText("%s" % feature.id())
+
+        fieldIdx = self.referencingFieldIndex()
+        widgetConfig = self.relation.referencingLayer().editorWidgetV2Config(fieldIdx)
+        self.relationWidgetWrapper = QgsEditorWidgetRegistry.instance().create("RelationReference",
+                                                                               self.relation.referencingLayer(),
+                                                                               fieldIdx,
+                                                                               widgetConfig,
+                                                                               self.relationReferenceWidget,
+                                                                               self,
+                                                                               self.editorContext)
+
+        self.relationWidgetWrapper.setEnabled(self.relation.referencingLayer().isEditable())
+        self.relationWidgetWrapper.setValue(feature[fieldIdx])
+        self.relationWidgetWrapper.valueChanged.connect(self.foreignKeyChanged)
+        # override field definition to allow map identification
+        self.relationReferenceWidget.setAllowMapIdentification(True)
+        self.relationReferenceWidget.setEmbedForm(False)
+
+        # update drawn link
+        self.highlightReferencingFeature()
+        self.drawLink()
+
+    def foreignKeyChanged(self, newKey):
+        if not self.relation.isValid() or not self.relation.referencingLayer().isEditable() or not self.referencingFeature.isValid():
+            self.drawLink()
+            return
+        if not self.relation.referencingLayer().editBuffer().changeAttributeValue(self.referencingFeature.id(), self.referencingFieldIndex(), newKey):
+            self.iface.messageBar().pushMessage("Link It", "Cannot change attribute value.", QgsMessageBar.CRITICAL)
+        self.drawLink()
+
+    def relationEditableChanged(self):
+        if self.relationWidgetWrapper is not None:
+            self.relationWidgetWrapper.setEnabled(self.relation.isValid() and self.relation.referencingLayer().isEditable())
+
+    def layerValueChangedOutside(self, fid, fieldIdx, value):
+        if not self.relation.isValid() or not self.referencingFeature.isValid() or self.relationWidgetWrapper is None:
+            return
+        # not the correct feature
+        if fid != self.referencingFeature.id():
+            return
+        # not the correct field
+        if fieldIdx != self.referencingFieldIndex():
+            return
+        # widget already has this value
+        if value == self.relationWidgetWrapper.value():
+            return
+        self.relationWidgetWrapper.valueChanged.disconnect(self.foreignKeyChanged)
+        self.relationWidgetWrapper.setValue(value)
+        self.relationWidgetWrapper.valueChanged.connect(self.foreignKeyChanged)
+
+    def referencingFieldIndex(self):
+        if not self.relation.isValid():
+            return -1
+        fieldName = self.relation.fieldPairs().keys()[0]
+        fieldIdx = self.relation.referencingLayer().fieldNameIndex(fieldName)
+        return fieldIdx
 
     @pyqtSlot(bool, name="on_drawButton_toggled")
-    def drawLink(self, dummy=None):
+    def drawLink(self):
+        self.settings.setValue("drawEnabled", self.drawButton.isChecked())
         self.linkRubber.reset()
-        ok, dstFeature = self.getFeature()
+        referencedFeature = self.relationReferenceWidget.referencedFeature()
 
-        if not ok or not self.drawButton.isChecked():
+        if not self.drawButton.isChecked() or not self.referencingFeature.isValid() or not referencedFeature.isValid() or not self.relation.isValid():
             return
 
-        # centroid of destination feature
-        p1 = centroid(dstFeature)
-        # centroid of source feature
-        srcFeature = QgsFeature()
-        srcId = dstFeature[self.destinationField]
-        if not srcId:
-            return
-        if self.sourceLayer.getFeatures(QgsFeatureRequest().setFilterFid(srcId)).nextFeature(srcFeature) is False:
-            return
-        p2 = centroid(srcFeature)
-        # point in middle
-        mp = QgsPoint((p1.x()+p2.x())/2, (p1.y()+p2.y())/2)
-        # distance between the two points
-        d = sqrt(p1.sqrDist(p2))
-        # orthogonal direction to segment p1-p2
-        az = (p1.azimuth(p2)+90)*pi/180
-        # create point distant to segment of offset of segment length, will be center of circular arc
-        # offset should be  0 < offset < 1
-        offset = 1/12
-        cp = QgsPoint(mp.x()+d*offset*sin(az),
-                      mp.y()+d*offset*cos(az))
-        # radius
-        r = d*sqrt(4*offset*offset+1)/2
-        # calculate start and end azimuth of circular arc
-        az1 = cp.azimuth(p1)
-        az2 = cp.azimuth(p2)
-        if az2 < az1:
-            az2 += 360
-        # draw arc
-        vx = [cp.x()+r*sin(az*pi/180) for az in floatrange(az1, az2, 5)]
-        vy = [cp.y()+r*cos(az*pi/180) for az in floatrange(az1, az2, 5)]
-        arc = [QgsPoint(vx[i], vy[i]) for i in range(len(vx))]
-        geom = QgsGeometry().fromPolyline(arc)
-        self.linkRubber.setToGeometry(geom, self.destinationLayer)
+        p1 = self.centroid(self.relation.referencedLayer(), referencedFeature)
+        p2 = self.centroid(self.relation.referencingLayer(), self.referencingFeature)
+        geom = arc(p1, p2)
+
+        self.linkRubber.setToGeometry(geom, None)
         self.linkRubber.setWidth(self.settings.value("rubberWidth"))
         self.linkRubber.setColor(self.settings.value("rubberColor"))
         self.linkRubber.setLineStyle(Qt.DashLine)
 
-        self.mapCanvas.refresh()
+    def centroid(self, layer, feature):
+        geom = feature.geometry()
+        if geom.type() == QGis.Line:
+            geom = geom.interpolate(geom.length()/2)
+        else:
+            geom = geom.centroid()
+        return self.iface.mapCanvas().mapSettings().layerToMapCoordinates(layer, geom.asPoint())
 
+    @pyqtSlot(name="on_highlightReferencingFeatureButton_clicked")
+    def highlightReferencingFeature(self):
+        self.deleteHighlight()
+        if not self.relation.isValid() or not self.referencingFeature.isValid():
+            return
+        
+        self.featureHighlight = QgsHighlight(self.iface.mapCanvas(), self.referencingFeature.geometry(), self.relation.referencingLayer())
+        settings = QSettings()
+        color = QColor( settings.value("/Map/highlight/color", QGis.DEFAULT_HIGHLIGHT_COLOR.name()))
+        alpha = int(settings.value("/Map/highlight/colorAlpha", QGis.DEFAULT_HIGHLIGHT_COLOR.alpha()))
+        buffer = float(settings.value("/Map/highlight/buffer", QGis.DEFAULT_HIGHLIGHT_BUFFER_MM))
+        minWidth = float(settings.value("/Map/highlight/minWidth", QGis.DEFAULT_HIGHLIGHT_MIN_WIDTH_MM))
+        
+        self.featureHighlight.setColor(color)
+        color.setAlpha(alpha)
+        self.featureHighlight.setFillColor(color)
+        self.featureHighlight.setBuffer(buffer)
+        self.featureHighlight.setMinWidth(minWidth)
+        self.featureHighlight.show()
 
+        timer = QTimer(self)
+        timer.setSingleShot(True)
+        timer.timeout.connect(self.deleteHighlight)
+        timer.start(3000)
 
-            
-            
-
+    def deleteHighlight(self):
+        if self.featureHighlight:
+            del self.featureHighlight
+        self.featureHighlight = None