+++ i18n/linkit_fr.ts
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE TS><TS version="2.0">
+<context>
+    <name>linker</name>
+    <message>
+        <location filename="ui_linker.py" line="123"/>
+        <source>Link It</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="ui_linker.py" line="124"/>
+        <source>Referenced</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="ui_linker.py" line="125"/>
+        <source>Referencing</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="ui_linker.py" line="126"/>
+        <source>Delete</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="ui_linker.py" line="127"/>
+        <source>Select</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="ui_linker.py" line="128"/>
+        <source>Relation</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="ui_linker.py" line="129"/>
+        <source>Draw connection arrows</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="ui_linker.py" line="130"/>
+        <source>Draw line</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+</TS>